@@ -4,16 +4,16 @@ version: 0.0.2
 homepage: https://github.com/deadblit/barras
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.17.0 <2.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.0.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  qrcode: ^1.0.4
-  torch_compat: ^1.0.2
-  flutter_beep: ^0.2.0
+  qrcode_forked: ^1.0.0
+  torch_controller: ^0.4.1
+  flutter_beep: ^1.0.0
 
 dev_dependencies:
   flutter_test:
